refactor: centralize hooks and rename modules
