+++ src/core/hooks/authentication/useLogin.ts
@@ -1,8 +1,9 @@
 import { AuthModel } from "@core/models/AuthModel";
-import { LoginInputs } from "../validation";
+import { LoginInputs } from "@modules/authentication/login/validation";
 import Khaos from "@shared/utilities/Khaos";
 import useKhaos from "@core/hooks/useKhaos";
 import Hook from "@core/interfaces/Hook";
+import TokenService from "@services/TokenService";
 
 const useLogin = (): Hook<AuthModel, LoginInputs> => {
   const { data, loading, error, fetch } = useKhaos<AuthModel>();
@@ -11,6 +12,8 @@ const useLogin = (): Hook<AuthModel, LoginInputs> => {
   const invoke = async (inputs: LoginInputs) => {
     khaos.setBody(inputs);
     await fetch(khaos);
+    
+    if (data) TokenService.save(data);
   };
 
   return { data, loading, error, invoke };
