+++ src/core/hooks/authentication/useReset.ts
@@ -1,5 +1,4 @@
 import { ResetPasswordModel } from "@core/models/ResetPasswordModel";
-/* import { ResetInputs } from "../validation"; */
 import Khaos from "@shared/utilities/Khaos";
 import useKhaos from "@core/hooks/useKhaos";
 import Hook from "@core/interfaces/Hook";
@@ -9,7 +8,7 @@ interface ResetInputs {
   token: string;
 };
 
-const useResetPassword = (): Hook<ResetPasswordModel, ResetInputs> => {
+const useReset = (): Hook<ResetPasswordModel, ResetInputs> => {
   const { data, loading, error, fetch } = useKhaos<ResetPasswordModel>();
   const khaos = new Khaos("/auth/reset-password").setHttpMethod("POST");
 
@@ -21,4 +20,4 @@ const useResetPassword = (): Hook<ResetPasswordModel, ResetInputs> => {
   return { data, loading, error, invoke };
 };
 
-export default useResetPassword;
+export default useReset;
